refactor: separate init into two parts (config - side effects)
